disable tests on osx and desktop (APIs not there/not supported)
